+++ Networking/Sockets/ListeningSocket.hpp
@@ -7,7 +7,7 @@ namespace NTW
 class ListeningSocket : public BindingSocket
 {
   public:
-    ListeningSocket(int domain, int service, int protocol, int port, u_long interface, int backlog);
+    ListeningSocket(int domain, int service, int protocol, uint16_t port, u_long interface, int backlog);
     virtual ~ListeningSocket();
     void start_listening();
